[Chore] #4 - Candy로 파일 생성했던 것 병합
