--- src/main/java/org/sopt/server/controller/CandyController.java
@@ -1,35 +1,0 @@
-package org.sopt.server.controller;
-
-import lombok.RequiredArgsConstructor;
-import org.sopt.server.global.response.ApiResponse;
-import org.sopt.server.global.response.SuccessType;
-import org.sopt.server.service.CandyService;
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PatchMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.ResponseStatus;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequiredArgsConstructor
-@RequestMapping("/candle")
-public class CandyController {
-
-    private final CandyService candyService;
-
-    @GetMapping
-    @ResponseStatus(HttpStatus.OK)
-    public ApiResponse<?> findById(@RequestParam Long candleId,
-                                   @RequestParam String cakeName) {
-        return ApiResponse.success(SuccessType.GET_CANDLE_SUCCESS, candyService.findById(candleId, cakeName));
-    }
-
-    @PatchMapping
-    @ResponseStatus(HttpStatus.NO_CONTENT)
-    public ApiResponse<?> updateDate(@RequestParam Long candleId) {
-        candyService.updateDate(candleId);
-        return ApiResponse.success(SuccessType.UPDATE_CANDLE_DATE_SUCCESS);
-    }
-}
